Refactored index.js into app.js and index.js
